Remove Avalanche, Fantom, and Beam support; add arcade credits
We decided to limit the number of networks we support and remove support for Avalanche, Fantom, and Beam. This change simplifies the hardhat configuration and other parts of the code. The hardhat.config.ts diff shows the simplification of the configuration file.

At the same time, we introduced the concept of credits to the arcade. Players now have a certain amount of credits they can use to play games. We added the necessary functions to deduct credits for each play and added a function to add credits to wallets. This information can be seen in BobaVerseArcade.sol diff.

This update also includes some minor fixes, such as updating Readme with the project name and new contracts.

Remove unused initialize file as it is no longer necessary for setting up the project environment.

Made updates in accordance with OpenZeppelin's changes in hardhat dependency, updated the typescript version and other dependencies in the package.json diff.

Please remember to update your .env file based on the updated .env.example.